fix: update return types to float for cutting length and area calculations; adjust order quantity type to float in purchase order

+++ utils/inventory/laser_cut_part.py
@@ -383,15 +383,15 @@ class LaserCutPart(InventoryItem):
         return "".join(f"{i + 1}. {category.name}: {self.get_category_quantity(category)}\n" for i, category in enumerate(self.categories))
 
     def calculate_machine_time_from_length(self, L: float) -> float:
-        return 0.00001 * L**2 + 0.00389 * L + 0.25198
+        return float(0.00001 * L**2 + 0.00389 * L + 0.25198)
 
     def calculate_machine_time_from_length_and_piercing_points(self, L: float, P: float) -> float:
-        return (
+        return float(
             0.00000 * L**3 + 0.00000 * L**2 * P + -0.00000 * L * P**2 + 0.00000 * P**3 + -0.00001 * L**2 + -0.00010 * L * P + 0.00011 * P**2 + 0.00816 * L + 0.01222 * P + 0.04308
         )
 
     def calculate_piercing_time(self, L: float, P: float) -> float:
-        return 0.00000 * L**2 * P + -0.00000 * L * P**2 + 0.00000 * P**3 + -0.00010 * L * P + 0.00011 * P**2 + 0.01222 * P
+        return float(0.00000 * L**2 * P + -0.00000 * L * P**2 + 0.00000 * P**3 + -0.00010 * L * P + 0.00011 * P**2 + 0.01222 * P)
 
     def calculate_weight(self) -> float:
         if pounds_per_square_foot := self.sheet_settings.get_pounds_per_square_foot(self.meta_data.material, self.meta_data.gauge):
